풀링 기능 추가, 어스 카운터 스킬 구현

+++ Assets/WooJeong/Scripts/BossMonster/Tree/BossData.cs
@@ -18,7 +18,7 @@ public class BossData : ScriptableObject
         Phase = 1;
     }
 
-    public void ReduceHp(int decrease)
+    public void ReduceHp(float decrease)
     {
         HP -= decrease;
         if (HP < 0)
